Add func_80056D64_57964 to graphics.h header
Fixed MISSING_HEADER_DECLARATION for func_80056D64_57964 by:
- Adding proper declaration to include/graphics.h with s32 for arg1
- Removing extern declarations from src/1E1BA0.c and src/594E0.c
- Updating implementation in src/graphics.c to use s32 for arg1

Using s32 instead of s16/u16 for the second parameter allows proper
type promotion from u16 array elements and matches the original assembly.

+++ src/graphics.c
@@ -261,7 +261,7 @@ void func_80056C6C_5786C(void *arg0, unsigned int arg1, s16 arg2) {
     }
 }
 
-void func_80056D64_57964(void *arg0, s16 arg1, s16 arg2, s16 arg3) {
+void func_80056D64_57964(void *arg0, s32 arg1, s16 arg2, s16 arg3) {
     RenderQueueItem *renderQueue;
     s32 index = gGraphicsManager->renderQueueCount;
     if (index < 0x20) {
